feat(coupon): 쿠폰 사용 처리 및 도메인 이벤트 핸들러 추가
- 쿠폰 사용 시 도메인 이벤트 발행 로직 구현
- 쿠폰 사용 상태 변경 및 이벤트 핸들링을 위한 CouponEventHandler 추가
- 주문 트랜잭션과 쿠폰 사용 처리를 이벤트로 분리하여 일관성 보장
