+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderFacade.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.IntStream;
 
-import org.springframework.context.ApplicationEventPublisher;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
@@ -17,7 +17,7 @@ import com.loopers.application.payment.PaymentFacade;
 import com.loopers.application.payment.PaymentInfo;
 import com.loopers.domain.coupon.CouponEntity;
 import com.loopers.domain.coupon.CouponService;
-import com.loopers.domain.coupon.event.CouponConsumeEvent;
+
 import com.loopers.domain.order.OrderEntity;
 import com.loopers.domain.order.OrderItemEntity;
 import com.loopers.domain.order.OrderService;
@@ -50,7 +50,7 @@ public class OrderFacade {
     private final PointService pointService;
     private final CouponService couponService;
     private final PaymentFacade paymentFacade;
-    private final ApplicationEventPublisher eventPublisher;
+
 
     /**
      * 주문 생성
@@ -116,11 +116,10 @@ public class OrderFacade {
         IntStream.range(0, orderableProducts.size())
                 .forEach(i -> productService.deductStock(orderableProducts.get(i), quantities.get(i)));
 
-        // 쿠폰 사용 이벤트 발행 (동기 처리)
-        List<Long> couponIds = coupons.stream()
+        // 쿠폰 사용 처리 (도메인 엔티티에서 이벤트 발행)
+        coupons.stream()
                 .filter(Objects::nonNull)
-                .map(CouponEntity::getId).toList();
-        eventPublisher.publishEvent(new CouponConsumeEvent(couponIds, user.getId(), creationResult.order().getId()));
+                .forEach(coupon -> coupon.useForOrder(creationResult.order().getId()));
 
         // 8. 주문 정보 반환
         return OrderFacadeDtos.OrderInfo.from(creationResult.order(), creationResult.orderItems());
@@ -188,12 +187,10 @@ public class OrderFacade {
         IntStream.range(0, orderableProducts.size())
                 .forEach(i -> productService.deductStock(orderableProducts.get(i), quantities.get(i)));
 
-
-        // 쿠폰 사용 이벤트 발행 (동기 처리)
-        List<Long> couponIds = coupons.stream()
+        // 쿠폰 사용 처리 (도메인 엔티티에서 이벤트 발행)
+        coupons.stream()
                 .filter(Objects::nonNull)
-                .map(CouponEntity::getId).toList();
-        eventPublisher.publishEvent(new CouponConsumeEvent(couponIds, user.getId(), creationResult.order().getId()));
+                .forEach(coupon -> coupon.useForOrder(creationResult.order().getId()));
 
         
         // 8. 주문 정보 반환 (PENDING 상태)
@@ -310,7 +307,6 @@ public class OrderFacade {
                 });
 
         // 4. 포인트는 차감하지 않았으므로 환불하지 않음
-
         return OrderFacadeDtos.OrderInfo.from(order, orderItems);
     }
