+++ apps/commerce-api/src/main/java/com/loopers/domain/coupon/event/CouponConsumeEvent.java
@@ -1,15 +1,13 @@
 package com.loopers.domain.coupon.event;
 
-import java.math.BigDecimal;
-import java.util.List;
+import com.loopers.domain.coupon.CouponEntity;
 
 /**
  * @author hyunjikoh
  * @since 2025. 12. 9.
  */
 public record CouponConsumeEvent (
-        List<Long> couponId,
-        Long userId,
+        CouponEntity coupon,
         Long orderId
 ){
 }
